Added exercise generation program comb.cc
The program comb.cc (courtesy of Howard Cheng) takes two files as input.
The first file contains all the problems for a section (one line per
problem) and the second contains the answers for each problem (each one
must occupy a single line). The output is a separate .tex file for each
problem compatible with the APEX formatting.